### 2018-04-29 #708 @whedon
[PRE REVIEW]: PyCM: Multi class confusion matrix library in python
**Submitting author:** @sepandhaghighi (<a href="http://orcid.org/0000-0001-9450-2375">Sepand Haghighi</a>)
**Repository:** <a href="https://github.com/sepandhaghighi/pycm" target ="_blank">https://github.com/sepandhaghighi/pycm</a>
**Version:** v0.8.1
