### 2018-04-29 0bef171
fix : chi-squared fixed in paper
