+++ paper/paper.bib
@@ -117,7 +117,7 @@
   number = {3},
   pages = {323},
   author = {Robert B. Davies},
-  title = {Algorithm {AS} 155: The Distribution of a Linear Combination of $\upchi$ 2 Random Variables},
+  title = {Algorithm {AS} 155: The Distribution of a Linear Combination of chi-squared Random Variables},
   journal = {Applied Statistics}
 }
 
