### 2018-05-03 f4730b9
Add license scan report and status
Signed-off-by: fossabot <badges@fossa.io>
