+++ paper/paper.md
@@ -28,7 +28,9 @@ bibliography: paper.bib
 
 # Summary			
 			
-In the field of machine learning, and specifically for statistical classification, a confusion matrix, also known as error matrix, is a specific table layout that allows visualization of the algorithm performance, and is mostly used in supervised learning. Each row of the matrix represents the instances in an actual class, while each column represents the instances in a predict class (or vice versa)[@Powers2011; @Sammut2010; @Shepperd2014; @Deng2016] .					
+In the field of machine learning, and specifically for statistical classification, a confusion matrix, also known as error matrix, is a specific table layout that allows visualization of the algorithm performance, and is mostly used in supervised learning. Each row of the matrix represents the instances in an actual class, while each column represents the instances in a predict class (or vice versa)[@Powers2011; @Sammut2010; @Shepperd2014; @Deng2016] .			
+
+![PyCM Block Diagram](../Otherfiles/block_diagram.jpg) 				
 
 PyCM is a multi-class confusion matrix library written in python that supports both input data vectors and direct matrix.			
 
@@ -40,7 +42,6 @@ We can categorize these statistics in 3 sections :
 2. Class Statistics			
 3. Overall Statistics
 
-![PyCM Block Diagram](../Otherfiles/block_diagram.jpg) 
 
 PyCM is also capable of generating report in HTML, CSV and .pycm formats.
 
