### 2018-05-03 #708 @sepandhaghighi
Hi @arfon

Could you please tell me what is the problem of image in `paper.md` ?

Thanks ;-)
