### 2018-05-03 93e19e5
fix : block_diagram.png added
