### 2018-05-04 #708 @arfon
> Oh fixed, after I changed size of the image to 600*662 in .png format.

Great thanks.

> But another question is about the image location??!!!

I think this is just the LaTeX engine being weird (image placement in LaTeX is notoriously peculiar) because the paper is very short. I suggest we fix this when we've been through review.
