### 2018-05-06 #708 @arokem
👋 Hey @rasbt: would you be willing to review this article for JOSS?
