### 2018-05-07 #708 @rasbt
Hi @arokem. Generally, I would be happy to do the review, but I am currently already involved in the mlpack review at JOSS, which will probably take a few weeks. As alternative reviewer with substantial expertise in that area, I would suggest @rhiever @NickleDave @jmschrei @jmwoloso
