### 2018-05-12 #729 @NickleDave
review in progress, would say "accept with minor revisions".
added issue about documenting API.
https://github.com/sepandhaghighi/pycm/issues/16
