### 2018-05-12 sepandhaghighi/pycm#17 @NickleDave
example usages
Hi again @sepandhaghighi, one more minor issue:

the examples in the README are great for demonstrating functionality, but I think examples demonstrating actual use cases would be helpful for people.

Again, there are many metrics included in the package; I would say pick three-five metrics that you think will be the ones most frequently used by all users, and then provide some more real-world examples of using pycm with those metrics.
E.g., using it in conjunction with scikit-learn to compare two or three different classifiers.

This would be a place that a simple visualization would really be helpful and make the concepts more intuitive.
I have often adapted this code from the scikit-learn docs to my own needs when I have to visualize a confusion matrix:
http://scikit-learn.org/stable/auto_examples/model_selection/plot_confusion_matrix.html#sphx-glr-auto-examples-model-selection-plot-confusion-matrix-py
If you do not want to add scikit-learn as a dependency, you might just add a `data` sub-package with outputs from some models you have trained, and the same goes for matplotlib or whatever you use to visualize, you could for now just have the plots as static .png files in your library.

Maybe there was a particular problem you were facing that inspired you to develop the package and you have true data + predictions from those models that you could include as example datasets?

I really think adding these sorts of examples will help you recruit users to whom it's not immediately obvious why they shouldn't just use the basic confusion_matrix functionality in scikit-learn.
