### 2018-05-12 #729 @NickleDave
added one other minor revision, asking for more specific example usages:
https://github.com/sepandhaghighi/pycm/issues/17
