### 2018-05-13 sepandhaghighi/pycm#16 @sepandhaghighi
Hi @NickleDave

Thanks for your point, I will fix this in next few hours.
