### 2018-05-13 sepandhaghighi/pycm#17 @sepandhaghighi
@NickleDave  Good point  ;-)
I will add real-world examples
