### 2018-05-13 5c0e7a2
fix : ConfusionMatrix docstring added fixed #16
