### 2018-05-13 62b4d48
doc : acceptable data types section added to README.md fixed #16
