### 2018-05-13 8e19a36
doc : jupyter notebook file updated fixed #16
