+++ Document/README.md
@@ -4,7 +4,7 @@
 
 ----------
 
-1. [Jupyter Notebook](Document.ipynb) (On Github)
+1. [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Document.ipynb) (On Github)
 2. [HTML](http://www.shaghighi.ir/pycm/doc/)
 
 
