### 2018-05-13 98582f6
fix : minor edit in COnfusionMatrix docstring
