### 2018-05-13 3633a9c
fix : minor edit in COnfusionMatrix docstring
