+++ pycm/pycm_obj.py
@@ -24,8 +24,7 @@ class ConfusionMatrix():
     [0, 1, 2]
     >>> cm.table
     {0: {0: 3, 1: 0, 2: 0}, 1: {0: 0, 1: 1, 2: 2}, 2: {0: 2, 1: 1, 2: 3}}
-    >>> cm2 = ConfusionMatrix(matrix={"Class1": {"Class1": 1, "Class2":2},
-    "Class2": {"Class1": 0, "Class2": 5}}) # Create CM Directly
+    >>> cm2 = ConfusionMatrix(matrix={"Class1": {"Class1": 1, "Class2":2},"Class2": {"Class1": 0, "Class2": 5}})
     >>> cm2
     pycm.ConfusionMatrix(classes: ['Class1', 'Class2'])
     '''
