### 2018-05-13 #729 @arokem
Thanks @NickleDave !

@sepandhaghighi : please let me know here when you've had a chance to address these comments!
