### 2018-05-13 sepandhaghighi/pycm#18 @NickleDave
DOI in paper.md
Hey @sepandhaghighi thanks for working on the issues.

One very minor one: I think the JOSS guidelines ask for you to put any relevant DOIs in the paper.md file.
See the example `paper.md`:
https://raw.githubusercontent.com/arfon/fidgit/master/paper/paper.md
```
In a somewhat meta fashion, Fidgit is publishing itself to figshare with DOI 'https://doi.org/10.6084/m9.figshare.828487' [@figshare_archive].
```
As far as I can tell, your library DOI is not in your `paper.md` currently.
Please add that.
