### 2018-05-13 sepandhaghighi/pycm#19 @NickleDave
statement of need in README *and* paper.md
One last issue @sepandhaghighi

You define confusion matrices in your README.md and paper.md but you don't state *who* the main users of the library will be, i.e., your 'target audience' as the JOSS guideliness state.

Please add some language specifying which users you think will find `pycm` most useful, e.g. `pycm is the swiss-army knife of confusion matrices, targeted mainly at data scientists that need a broad array of metrics for predictive models`

Make sure to add this to both the README.md and the paper.md.
Thanks! That's the last "major-minor" revision :)
