### 2018-05-13 #729 @sepandhaghighi
Hi

@NickleDave Thanks again for review ;-)

@arokem I fixed first issue and working on second
