### 2018-05-13 #729 @NickleDave
added two other issues, after those are closed, this should be ready to accept
https://github.com/sepandhaghighi/pycm/issues/19
https://github.com/sepandhaghighi/pycm/issues/18
