### 2018-05-14 f66aa6b
doc : statement of need added to paper fixed #19
