### 2018-05-14 ad25a20
doc : README.md and Notebook files updated fixed #19
