+++ setup.py
@@ -11,14 +11,15 @@ setup(
     name='pycm',
     packages=['pycm'],
     version='0.8.1',
-    description='Multiclass confusion matrix library in python',
+    description='Multi-class confusion matrix library in python',
     long_description='''
-   In the field of machine learning and specifically the problem of statistical classification,
-   a confusion matrix, also known as an error matrix, is a specific table layout that allows visualization of
-   the performance of an algorithm, typically a supervised learning one (in unsupervised learning it is usually
-   called a matching matrix). Each row of the matrix represents the instances in a predicted class while each column
-   represents the instances in an actual class (or vice versa) pycm(python confusion matrix) is a multi-class
-   confusion matrix library in python.''',
+   PyCM is a multi-class confusion matrix library written in python that
+   supports both input data vectors and direct matrix, and a proper tool for
+   post-classification model evaluation that supports most classes and overall
+   statistics parameters.
+   PyCM is the swiss-army knife of confusion matrices, targeted mainly at
+   data scientists that need a broad array of metrics for predictive models
+   and an accurate evaluation of large variety of classifiers.''',
     author='Sepand Haghighi',
     author_email='sepand@qpage.ir',
     url='https://github.com/sepandhaghighi/pycm',
