### 2018-05-14 064142e
doc : PyCM reference added to paper fixed #18
