+++ paper/paper.bib
@@ -205,3 +205,13 @@
   title = {An improved method to construct basic probability assignment based on the confusion matrix for classification problem},
   journal = {Information Sciences}
 }
+
+@misc{Haghighi2018,
+  doi = {10.5281/zenodo.1157173},
+  author = {{Sepand Haghighi} and Jasemi,  Masoomeh and {Shaahin Hessabi}},
+  keywords = {Machine Learning,  statistics,  confusion-matrix,  statistical-analysis,  matrices,  matrix},
+  title = {Pycm : Multi Class Confusion Matrix Library In Python},
+  pages = {--},
+  publisher = {Zenodo},
+  year = {2018}
+}
