### 2018-05-14 83e58f6
doc : add alireza information to paper file
