+++ paper/paper.md
@@ -18,9 +18,15 @@ authors:
  - name: Shaahin Hessabi
    orcid: 0000-0003-3193-2567
    affiliation: 1
+ - name: Alireza Zolanvari
+   orcid: 0000-0003-2367-8343
+   affiliation: 2
 affiliations:
  - name: Sharif University of Technology
    index: 1
+ - name: Amirkabir University of Technology
+   index: 2
+
 date: 29 April 2018
 bibliography: paper.bib
 ---
