### 2018-05-15 c5ba6c6
doc : Example file updated fixed #17
