+++ Document/README.md
@@ -1,11 +1,20 @@
-# PYCM Document
+# PyCM Document
 
 ----------
 
 ----------
 
-1. [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Document.ipynb) (On Github)
-2. [HTML](http://www.shaghighi.ir/pycm/doc/)
+## Document	
+
+- [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Document.ipynb)
+- [HTML](http://www.shaghighi.ir/pycm/doc/)				
+
+
+## Example	
+
+- [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Document.ipynb)
+- [HTML](http://www.shaghighi.ir/pycm/doc/Example.html)
+
 
 
 																												
