### 2018-05-15 #729 @sepandhaghighi
@NickleDave @arokem
Thank you for your time
The issues have been fixed. ;-)
