### 2018-05-15 a561385
doc : blockdiagram removed from paper
