### 2018-05-15 c30c473
fix : minor edit in paper
