### 2018-05-15 ece6b48
doc : minor edit in paper
