### 2018-05-15 f8ca2d2
doc : minor edit in paper reference style
