+++ paper/paper.bib
@@ -208,7 +208,7 @@
 
 @misc{Haghighi2018,
   doi = {10.5281/zenodo.1157173},
-  author = {{Sepand Haghighi} and Jasemi,  Masoomeh and {Shaahin Hessabi}},
+  author = {Sepand Haghighi and Masoomeh Jasemi and Shaahin Hessabi},
   keywords = {Machine Learning,  statistics,  confusion-matrix,  statistical-analysis,  matrices,  matrix},
   title = {Pycm : Multi Class Confusion Matrix Library In Python},
   pages = {--},
