### 2018-05-15 8c430b7
fix : minor edit in text
