+++ Document/README.md
@@ -12,7 +12,7 @@
 
 ## Example	
 
-- [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Document.ipynb)
+- [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Example.ipynb)
 - [HTML](http://www.shaghighi.ir/pycm/doc/Example.html)
 
 
