### 2018-05-16 32c4898
fix : minor edit in paper and readme
