### 2018-05-16 20853dd
doc : marplotlib example added to Example2 notebook
