+++ Document/README.md
@@ -10,10 +10,17 @@
 - [HTML](http://www.shaghighi.ir/pycm/doc/)				
 
 
-## Example	
+## Example-1 (Comparison of three different classifiers)	
 
-- [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Example.ipynb)
-- [HTML](http://www.shaghighi.ir/pycm/doc/Example.html)
+- [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Example1.ipynb)
+- [HTML](http://www.shaghighi.ir/pycm/doc/Example1.html)
+
+## Example-2 (How to plot via matplotlib)
+
+- [Jupyter Notebook](https://nbviewer.jupyter.org/github/sepandhaghighi/pycm/blob/master/Document/Example2.ipynb)
+- [HTML](http://www.shaghighi.ir/pycm/doc/Example2.html)
+
+
 
 
 
