### 2018-05-16 1d2b380
fix : minor edit in setup file
