+++ AUTHORS.md
@@ -4,3 +4,4 @@
 - Sepand Haghighi - Sharif University Of Technology ([shaghighi@ce.sharif.edu](mailto:shaghighi@ce.sharif.edu))
 - Masoomeh Jasemi - Sharif University Of Technology ([jasemi@ce.sharif.edu](mailto:jasemi@ce.sharif.edu))
 - Shaahin Hessabi - Sharif University Of Technology ([hessabi@sharif.edu](mailto:hessabi@sharif.edu))
+- Alireza Zolanvari  - Amirkabir University of Technology ([alirezazolanvari@aut.ac.ir](mailto:alirezazolanvari@aut.ac.ir))
