### 2018-05-17 9bd350f
fix : minor bug in html report fixed
