+++ pycm/pycm_output.py
@@ -56,10 +56,11 @@ def html_table(classes, table):
     result += '<tr align="center">\n<td></td>\n'
     part_2 = ""
     for i in classes:
-        result += '<td style="border:1px solid black;padding:10px;">' + \
+        result += '<td style="border:1px solid ' \
+                  'black;padding:10px;height:75px;width:75px;">' + \
             str(i) + '</td>\n'
         part_2 += '<tr align="center">\n'
-        part_2 += '<td style="border:1px solid black;padding:10px;">' + \
+        part_2 += '<td style="border:1px solid black;padding:10px;height:75px;width:75px;">' + \
             str(i) + '</td>\n'
         for j in classes:
             item = table[i][j]
@@ -67,7 +68,7 @@ def html_table(classes, table):
             back_color = html_table_color(table[i], item)
             if back_color < 128:
                 color = "white"
-            part_2 += '<td style="background-color:	rgb({0},{0},{0});color:{1};padding:10px">'.format(
+            part_2 += '<td style="background-color:	rgb({0},{0},{0});color:{1};padding:10px;height:75px;width:75px;">'.format(
                 str(back_color), color) + str(item) + '</td>\n'
         part_2 += "</tr>\n"
     result += '</tr>\n'
