### 2018-05-17 162a7b3
fix : notebooks updated with html bug fixed
