### 2018-05-20 #729 @arokem
@sepandhaghighi : Great! I believe your paper is ready to be accepted.

At this point, please create a new release of the software and create an archive for it (e.g., using https://zenodo.org/). Once you have that, please reply here with the DOI, so we can set that as the archive to be included in the paper.
