+++ setup.py
@@ -28,14 +28,14 @@ def read_description():
 setup(
     name='pycm',
     packages=['pycm'],
-    version='0.8.1',
+    version='0.8.5',
     description='Multi-class confusion matrix library in Python',
     long_description=read_description(),
     long_description_content_type='text/markdown',
     author='Sepand Haghighi',
     author_email='sepand@qpage.ir',
     url='https://github.com/sepandhaghighi/pycm',
-    download_url='https://github.com/sepandhaghighi/pycm/tarball/v0.8.1',
+    download_url='https://github.com/sepandhaghighi/pycm/tarball/v0.8.5',
     keywords="confusion-matrix python3 python machine_learning ML",
     project_urls={
         'Webpage': 'http://pycm.shaghighi.ir',
