### 2018-05-20 80227f0
doc : README file links changed to global
