### 2018-05-20 a0c4b6a
fix : testcases fixed
