+++ pycm/pycm_param.py
@@ -4,13 +4,13 @@ VERSION = "0.8.5"
 
 
 OVERVIEW = '''
-   PyCM is a multi-class confusion matrix library written in Python that
-   supports both input data vectors and direct matrix, and a proper tool for
-   post-classification model evaluation that supports most classes and overall
-   statistics parameters.
-   PyCM is the swiss-army knife of confusion matrices, targeted mainly at
-   data scientists that need a broad array of metrics for predictive models
-   and an accurate evaluation of large variety of classifiers.
+PyCM is a multi-class confusion matrix library written in Python that
+supports both input data vectors and direct matrix, and a proper tool for
+post-classification model evaluation that supports most classes and overall
+statistics parameters.
+PyCM is the swiss-army knife of confusion matrices, targeted mainly at
+data scientists that need a broad array of metrics for predictive models
+and an accurate evaluation of large variety of classifiers.
 '''
 
 PARAMS_DESCRIPTION = {
