### 2018-05-20 7b55530
fix : minor edit in PyCM spell
