+++ pycm/pycm_output.py
@@ -17,7 +17,8 @@ def html_init(name):
     result += "<title>" + str(name) + "</title>\n"
     result += "</head>\n"
     result += "<body>\n"
-    result += '<h1 style="border-bottom:1px solid black;text-align:center;">PYCM Report</h1>'
+    result += '<h1 style="border-bottom:1px solid ' \
+              'black;text-align:center;">PyCM Report</h1>'
     return result
 
 
@@ -152,7 +153,8 @@ def html_end(version):
     '''
     result = "</body>\n"
     result += '<p style="text-align:center;position:absoloute;border-top:1px solid black;">Generated By ' \
-        '<a href="http://pycm.shaghighi.ir" style="text-decoration:none;color:red;">PYCM</a> Version ' + version + '</p>\n'
+        '<a href="http://pycm.shaghighi.ir" ' \
+              'style="text-decoration:none;color:red;">PyCM</a> Version ' + version + '</p>\n'
     result += "</html>"
     return result
 
