### 2018-05-20 11e2370
fix : minor bug in paper.md fixed
