### 2018-05-24 08f38ac
Update pytest from 3.5.1 to 3.6.0
