+++ requirements.txt
@@ -1,7 +1,7 @@
 art==1.0
 numpy==1.14.3
 codecov==2.0.15
-pytest==3.5.1
+pytest==3.6.0
 pytest-cov==2.5.1
 
 
