+++ requirements.txt
@@ -1,4 +1,4 @@
-art==1.0
+art==1.1
 numpy==1.14.3
 codecov==2.0.15
 pytest==3.5.1
