### 2018-05-30 #729 @NickleDave
Nice work @sepandhaghighi
Have you thought about sharing PyCM on the scikit-learn mailing list (scikit-learn@python.org )? I bet a lot of people there would be interested
